make it work in standard mode (instead of quirks mode)
In standard mode, you can't have unitless style.left/right..

Example file was in quirks mode, was misleading. Beofre this PR,
slideways is likely to work only on pages that are in quirksmode, thus
not a lot of pages.

https://developer.mozilla.org/en-US/docs/Mozilla_Quirks_Mode_Behavior
has a reference to the 'px' only added in quirksmode automatically.

Tested in chrome latest and IE9, I guess it's ok everywhere else.

This PR is similar to #1 but less noisy